Refactor the modules and files accoording to the discussion, fixed the imports, test would fail though.
