+++ pipeline_tools/pipelines/cellranger/cellranger.py
@@ -1,0 +1,66 @@
+from pipeline_tools.shared import metadata_utils
+from pipeline_tools.shared import tenx_utils
+from pipeline_tools.shared.http_requests import HttpRequests
+
+
+def get_cellranger_input_files(uuid, version, dss_url):
+    """ Get inputs for cellranger count workflow
+
+    Args:
+        uuid (str): the bundle uuid
+        version (str): the bundle version
+        dss_url (str): the DCP Data Storage Service
+
+    Returns:
+        None
+
+    Raises:
+        optimus_utils.LaneMissingFileError: if any fastqs are missing
+    """
+    # Get bundle manifest
+    print('Getting bundle manifest for id {0}, version {1}'.format(uuid, version))
+    primary_bundle = metadata_utils.get_bundle_metadata(
+        uuid=uuid, version=version, dss_url=dss_url, http_requests=HttpRequests()
+    )
+
+    sample_id = metadata_utils.get_sample_id(primary_bundle)
+    print('Writing sample ID to sample_id.txt')
+    with open('sample_id.txt', 'w') as f:
+        f.write('{0}'.format(sample_id))
+
+    total_estimated_cells = metadata_utils.get_expected_cell_count(primary_bundle)
+    print('Writing total estimated cells to expect_cells.txt')
+    with open('expect_cells.txt', 'w') as f:
+        f.write('{0}'.format(total_estimated_cells))
+
+    # Parse inputs from metadata
+    print('Gathering fastq inputs')
+    fastq_files = [
+        f for f in primary_bundle.files.values() if f.file_format == 'fastq.gz'
+    ]
+    lane_to_fastqs = tenx_utils.create_fastq_dict(fastq_files)
+
+    # Stop if any fastqs are missing
+    tenx_utils.validate_lanes(lane_to_fastqs)
+
+    read_indices = {'read1': 'R1', 'read2': 'R2', 'index1': 'I1'}
+    fastq_urls = []
+    fastq_names = []
+
+    for lane, reads in lane_to_fastqs.items():
+        for read_index, url in reads.items():
+            new_file_name = '{}_S1_L00{}_{}_001.fastq.gz'.format(
+                sample_id, str(lane), read_indices[read_index]
+            )
+            fastq_names.append(new_file_name)
+            fastq_urls.append(url)
+
+    with open('fastqs.txt', 'w') as f:
+        for url in fastq_urls:
+            f.write(url + '\n')
+
+    with open('fastq_names.txt', 'w') as f:
+        for name in fastq_names:
+            f.write(name + '\n')
+
+    print('Finished writing files')
